Add Portuguese localization and update dependencies
- Created a new localization file for Portuguese (pt) with translations for various app strings.
- Updated Dutch (nl) and Portuguese (pt) ARB files to include new fields for purchase value and hints in the purchase calculator.
- Changed import paths for localization files in main.dart and locale_provider.dart.
- Adjusted theme text styles in theme_provider.dart for better readability.
- Updated pubspec.lock and pubspec.yaml to reflect new package versions and dependencies.

+++ lib/features/onboarding/views/onboarding.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:dots_indicator/dots_indicator.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:card_cambio/l10n/app_localizations.dart';
 
 class Onboarding extends StatefulWidget {
   final VoidCallback onFinish;
@@ -91,7 +91,7 @@ class OnboardingState extends State<Onboarding> {
               children: [
                 DotsIndicator(
                   dotsCount: 3,
-                  position: _currentIndex,
+                  position: _currentIndex.toDouble(),
                   decorator: DotsDecorator(
                     activeColor: Theme.of(context).primaryColor,
                   ),
